feat: add a response `model`, `schema` and `mapper` per endpoint
EX-6078
