+++ packages/x-adapter-platform/src/__tests__/__fixtures__/identifier-results.response.ts
@@ -1,4 +1,4 @@
-export const skuSearchResponse = {
+export const identifierResultsResponse = {
   results: [
     {
       id: '12345-U',
