+++ packages/x-adapter-platform/src/mappers/response/__tests__/next-queries-response.mapper.spec.ts
@@ -1,5 +1,5 @@
-import { PlatformNextQueriesResponse } from '../../../types/response.types';
-import { nextQueriesResponseMapper } from '../../response/next-queries-response.mapper';
+import { PlatformNextQueriesResponse } from '../../../types/responses/next-queries-response.model';
+import { nextQueriesResponseMapper } from '../next-queries-response.mapper';
 
 describe('next queries response mapper', () => {
   it('should map the next queries', () => {
