+++ packages/x-adapter-platform/src/mappers/response/__tests__/related-tags-response.mapper.spec.ts
@@ -1,5 +1,5 @@
-import { PlatformRelatedTagsResponse } from '../../../types/response.types';
-import { relatedTagsResponseMapper } from '../../response/related-tags-response.mapper';
+import { PlatformRelatedTagsResponse } from '../../../types/responses/related-tags-response.model';
+import { relatedTagsResponseMapper } from '../related-tags-response.mapper';
 
 describe('related tags response mapper', () => {
   it('should map the related tags', () => {
