+++ packages/x-adapter-platform/src/mappers/response/__tests__/search-response.mapper.spec.ts
@@ -1,5 +1,5 @@
-import { searchResponseMapper } from '../../response/search-response.mapper';
-import { PlatformSearchResponse } from '../../../types/response.types';
+import { searchResponseMapper } from '../search-response.mapper';
+import { PlatformSearchResponse } from '../../../types/responses/search-response.model';
 import { platformSearchResponse } from '../../../__tests__/__fixtures__/platform-search.response';
 import { searchResponse } from '../../../__tests__/__fixtures__/search.response';
