+++ packages/x-adapter-platform/src/mappers/response/empathize-response.mapper.ts
@@ -1,16 +1,9 @@
-import { createMutableSchema, Schema, schemaMapperFactory } from '@empathyco/x-adapter-next';
-import { EmpathizeResponse, PlatformEmpathizeResponse } from '../../types/response.types';
-import { suggestionMutableSchema } from '../../schemas/suggestion.schema';
-
-export const empathizeResponseSchema: Schema<PlatformEmpathizeResponse, EmpathizeResponse> = {
-  suggestions: {
-    $path: 'topTrends.content',
-    $subSchema: suggestionMutableSchema
-  },
-  spellcheck: 'topTrends.spellcheck'
-};
-
-export const empathizeResponseMutableSchema = createMutableSchema(empathizeResponseSchema);
+import { schemaMapperFactory } from '@empathyco/x-adapter-next';
+import {
+  EmpathizeResponse,
+  PlatformEmpathizeResponse
+} from '../../types/responses/empathize-response.model';
+import { empathizeResponseMutableSchema } from '../../schemas/response/empathize-response.schema';
 
 export const empathizeResponseMapper = schemaMapperFactory<
   PlatformEmpathizeResponse,
