+++ packages/x-adapter-platform/src/mappers/response/index.ts
@@ -1,6 +1,6 @@
-export * from './search-response.mapper';
 export * from './empathize-response.mapper';
-export * from './top-clicked-response.mapper';
+export * from './identifier-results-response.mapper';
 export * from './next-queries-response.mapper';
 export * from './related-tags-response.mapper';
-export * from './sku-search-response.mapper';
+export * from './search-response.mapper';
+export * from './top-clicked-response.mapper';
