+++ packages/x-adapter-platform/src/mappers/response/next-queries-response.mapper.ts
@@ -1,17 +1,9 @@
-import { createMutableSchema, Schema, schemaMapperFactory } from '@empathyco/x-adapter-next';
-import { NextQueriesResponse, PlatformNextQueriesResponse } from '../../types/response.types';
-import { nextQueryMutableSchema } from '../../schemas/next-query.schema';
-
-export const nextQueriesResponseSchema: Schema<PlatformNextQueriesResponse, NextQueriesResponse> = {
-  nextQueries: {
-    $path: 'data.nextqueries',
-    $subSchema: nextQueryMutableSchema
-  }
-};
-
-export const nextQueriesReponseMutableSchema = createMutableSchema(nextQueriesResponseSchema);
+import { schemaMapperFactory } from '@empathyco/x-adapter-next';
+import { NextQueriesResponse } from '@empathyco/x-types';
+import { nextQueriesResponseMutableSchema } from '../../schemas/response';
+import { PlatformNextQueriesResponse } from '../../types/responses/next-queries-response.model';
 
 export const nextQueriesResponseMapper = schemaMapperFactory<
   PlatformNextQueriesResponse,
   NextQueriesResponse
->(nextQueriesReponseMutableSchema);
+>(nextQueriesResponseMutableSchema);
