+++ packages/x-adapter-platform/src/schemas/index.ts
@@ -1,4 +1,5 @@
 export * from './request';
+export * from './response';
 export * from './result.schema';
 export * from './banner.schema';
 export * from './facet.schema';
