+++ packages/x-adapter-platform/src/schemas/response/index.ts
@@ -1,0 +1,6 @@
+export * from './empathize-response.schema';
+export * from './identifier-results-response.schema';
+export * from './next-queries-response.schema';
+export * from './related-tags-response.schema';
+export * from './search-response.schema';
+export * from './top-clicked-response.schema';
