+++ packages/x-adapter-platform/src/types/responses/identifier-results-response.model.ts
@@ -1,0 +1,10 @@
+import { PlatformResult } from '../models.types';
+
+export interface PlatformIdentifierResultsResponse {
+  catalog: {
+    content: PlatformResult[];
+    tagging: {
+      query: string;
+    };
+  };
+}
