+++ packages/x-adapter-platform/src/types/responses/index.ts
@@ -1,0 +1,6 @@
+export * from './empathize-response.model';
+export * from './identifier-results-response.model';
+export * from './next-queries-response.model';
+export * from './related-tags-response.model';
+export * from './search-response.model';
+export * from './top-clicked-response.model';
