+++ packages/x-adapter-platform/src/types/responses/next-queries-response.model.ts
@@ -1,0 +1,7 @@
+import { PlatformNextQuery } from '../models.types';
+
+export interface PlatformNextQueriesResponse {
+  data: {
+    nextqueries: PlatformNextQuery[];
+  };
+}
