+++ packages/x-adapter-platform/src/types/responses/related-tags-response.model.ts
@@ -1,0 +1,8 @@
+import { PlatformRelatedTag } from '../models.types';
+
+export interface PlatformRelatedTagsResponse {
+  data: {
+    relatedtags: PlatformRelatedTag[];
+  };
+  status: number;
+}
