+++ packages/x-adapter-platform/src/types/responses/search-response.model.ts
@@ -1,0 +1,14 @@
+import {
+  Contentable,
+  PlatformBanner,
+  PlatformCatalog,
+  PlatformPromoted,
+  PlatformRedirection
+} from '../models.types';
+
+export interface PlatformSearchResponse {
+  banner: Contentable<PlatformBanner>;
+  catalog: PlatformCatalog;
+  direct: Contentable<PlatformRedirection>;
+  promoted: Contentable<PlatformPromoted>;
+}
