+++ packages/x-adapter-platform/src/types/responses/top-clicked-response.model.ts
@@ -1,0 +1,13 @@
+import { Result } from '@empathyco/x-types';
+import { PlatformResult } from '../models.types';
+
+export interface PlatformTopClickedResponse {
+  topclicked: {
+    content: PlatformResult[];
+    numFound: number;
+  };
+}
+
+export interface TopClickedResponse {
+  results: Result[];
+}
